requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.8
decorator==4.4.2
exceptiongroup==1.1.1
faiss-cpu==1.7.4
filelock==3.12.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
Jinja2==3.1.2
jsonschema==4.17.3
langchain==0.0.207
langchainplus-sdk==0.0.16
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.57.1
numexpr==2.8.4
numpy==1.24.4
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
proglog==0.1.10
protobuf==4.23.3
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
selenium==4.10.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.16
srt==3.5.3
streamlit==1.23.1
streamlit-lottie==0.0.5
sympy==1.12
tenacity==8.2.2
tiktoken==0.3.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.6
validators==0.20.0
vosk==0.3.44
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
